Add python_2_unicode_compatible decorator to models
This commit adds the python_2_unicode_compatible decorator to the
models defined in models.py.  This provides a __unicode__ method
under Python 2 and does nothing in Python 3.  The __unicode__
method is not available in Python 3.

This is being added as we have unicode data in our Country table which
causes the template to fail to render with no __unicode__ method
defined.

See [1] for more information.

1. https://docs.djangoproject.com/en/1.9/topics/python3/#str-and-unicode-methods