Add activate method to job and expand expire method
This commit adds an activate method to the job model and updates the
expire method so that already expired jobs cannot be expired again.
We also remove a duplicate jobs_activate method from
job_board/views.py.

Lastly, this commit adds a number of new tests to properly test the
activate and expire methods.