Only show active jobs when browsing category
Currently, all jobs (include those expired) are displayed when view
a category.  This commit updates the view to only select those that
are active.

Note that switch from using get_list_or_404 here, which means we no
longer get a 404 and requires a test update.  It doesn't seem logical
to return a 404 on the categories_show view when no jobs in that
category exist as that will wreak havoc w/ search engines, etc.