Add SiteConfig model
This commit adds a new model called SiteConfig, which is basically
needed to store configuration for individual sites stored in the Site
model.  The Site model is created when the sites framework is used, so
it's not straight-forward to simply add new attributes to this model.

Additionally, we create a signal that creates a SiteConfig object when
a new site is added.  This is needed so we can ensure that a site has
a correct default SiteConfig when it is added.

As sites now have a custom expiration value (default 30), we also update
the expire command to cycle through sites and then expire jobs that are
older than the site's expire_after value.

Lastly, we add some tests to ensure that this signal is getting fired
correctly when sites are added.