Add customisable admin email address
Currently, we assume a site's admin address is admin@site.domain, which
is a poor assumption.  This commit makes this address customisable, but
defaults to admin@site.domain when the SiteConfig instance is created.

We also fix a bug in the expire command where we're filtering
unnecessarily, and update a test to also ensure that the admin email is
getting created correctly by default.