Set correct page titles
Currently, the base template has a title tag that says 'Categories',
and this is being displayed on every page.  This commit updates
views.py so that actually set the title/header from within the view
and then updates the templates to use {{ title }} instead of using
block tags.

Initially, the problem that we were running into was that we wanted to
use {% block header %}{% endblock %} inside the title tag, but this was
causing the template to error out since we're already using the header
block tag.  Rather than try to circumvent this, it made more sense to
just set the title (header) in views.py which ultimatley means we can
make the page titles customisable by site.  We can then also separate
out title/header should that be necessary.

Closes #35