[WIP] Remove usage of CurrentSiteManager
This commit removes the usage of CurrentSiteManager as we were running
into some issues with it.  Whilst it made filtering queries much easier
using on_site, it required you to have a custom settings.py for each
site with a unique SITE_ID in it for the site in question.  This adds a
fair bit of overhead when deploying new sites as you then need a
site-specific settings.yml file, a site-specific wsgi file, and a
web server configuration that points the specific site at the
site-specific wsgi file.CurrentSiteManager.  Also, while not 100%
confirmed, it appeared that /admin was then site-specific, meaning you
didn't have a global /admin to manage all sites in one pane.

Since CurrentSiteManager adds site to the request object, we needed to
create a custom context processor (get_site) which gives us access to
the site in all templates.  This is most importantly needed for
base.html, since we do not have a view-specific context dict where we
can append the site to.

Lastly, we update views.py to restrict categories and companies to the
site in question when adding/editing jobs.

Closes #41