Hide remote field on job when job board is remote
This commit udpates jobs.py view to remove the remote field when the
job board from the forms when the job board is remote.  The remote
field is then set when the job is saved.

Additionally, we display on individual job posts whether the job
supports remote working.  This should eventually be replaced by an
icon or something a bit more visible.