Add protocol to SiteConfig
This commit updates SiteConfig by adding a protocol field, which is
limited to either http or https.  This protocol is then used to create
URLs in templates.  As we have e-mail templates that are triggered via
management commands, we cannot simply use request.is_secure to
determine if we should use http or https.

We also remove the hard-coded reference to Matt in
job_board/templates/job_board/emails/expired.txt.

Closes #92