Job admin doesn't show when user is not an admin
We have a bug in the jobs_show.html template which show the job admin
panel when a user owns the job or if the user is an admin.  Currently,
we have a typo ensuring that user.id == job.id, but this should be
user.id == job.user_id.  This typo prevents non-admins from being
seeing the job admin panel.