Do not require city/state on job submission
Currently, if you post a job on a non-remote job board and select
the remote checkbox, it still requires you to enter a city and state.
This seems unnecessary as most 100% remote jobs are not constrained by
a city and state.  This commit removes the requirement to enter a city
and state on job submission, and also adds some basic jQuery to hide
city and state fields if remote is checked.

This is the first bit of JavaScript/jQuery being added to this project,
and as such it's not clear at this point in time how and where this
code should live.  We will figure this out and clean up as we add more
of this code to this code base.