Send e-mail when job is posted
This commit updates the jobs_new method in views/jobs to send the site
administrator an e-mail when a job is posted. This is necessary for
the administrator to activate the job when a new post comes in.

Note that we could have used a signal here, however we probably don't
want to always send an e-mail ie. when bulk-importing jobs via a
script etc.

Closes #116