Fix bug in companies_show template
This commit fixes a bug in companies_show where we do not correctly
show the company edit link if the company owner is the current user
(we are checking company.id rather than company.user.id).

Note that we do actually have a test in place for this, however the
test was passing since in this instance company.id did match
user.id.

We also update jobs_show to use job.user.id instead of job.user_id.
This is simply done for consistency.