Add MailChimp form
This commit adds a new url, view, form, and partial template for
allowing signups to a Mailchimp mailing list.  This also adds new
fields to the SiteConfig model for Mailchimp username, API key, and
list ID, and updates the jobs_index.html template to include the
partial template.

Note that we'll need to do some refinement on the subscribe view in
misc.py (specifically handling different error scenarios), but for now
this should allow us to start collecting e-mail addresses when the
necessary Mailchimp fields are populated.