Initial pass at URL slugs
This commit adds slugs to the following views:

* jobs_show
* companies_show
* categories_show

The old non-slug URLs are still present, and if someone hits a resource
on any of them it will redirect 301 to the slug.

We also override get_absolute_url() in the job, company, and category
models which means we do not need to do a reverse lookup on ID and
slug.  This is a fair bit less code and generally looks cleaner.

All applicable tests now use get_absolute_url() and a test for each
view has been added to ensure the redirect 301 happens correctly.