Add company on job post page (#144)
This commit removes the 'Add Company' link in the header and adds a
from to add a company in the jobs_new.html page.  This form uses an
ajax request so that we can add a company without page fresh, and once
the job is added the company is automatically selected in the dropdown
and the company add form disappears.

Note that companies_add view in companies.py should still handle adding
a company separately, so that the old template will still work.
Additionally, _errors.html template is updated to display all errors,
not just non_field_errors.  This is because companies_add view renders
this template, so in its previous form all non non_field_errors would
be lost.

Lastly, I don't do JavaScript, so apologies if any of this jQuery
committed here is an abomination.

Closes #72