Category RSS feeds (#159)
This commit adds in an initial RSS feed for job categories.  At
present, up to 30 max will be returned in the feed.  This may need to
be made customisable at a later date.

We also add font awesome fonts via CDN.  We may want to look at pulling
these into source control in the near future.

Closes #156