HTMLize feed description
Currently, the feed description is output as markdown.  This commit
converts it to HTML so that RSS readers render the description as HTML.

Additionally, we update the base template to display a link to an RSS
link when a URL is passed in from the view.  At present we only do this
from the categories_show view.