Add display_lists command
This adds a display_lists command which is helpful when configuring
Tramcar to send a weekly MailChimp mailshot for a given site.  The
output of the command includes a list of MailChimp list IDs and list
names, which can then be used to populate `mailchimp_list_id` for that
site.

Obviously `mailchimp_username` and `mailchimp_api_key` needs to have
been set in the admin panel for the site in question.

Example:

```
(.venv) $ python manage.py display_lists <site_domain>
```